requirements.txt
python-telegram-bot[job-queue]==21.7
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
